remove song speed and track length from song modifications
they can be changed in the track, and it will be confusing
if you save after encountering a change track-length or
change song-speed.  (unless one adds more variables to keep
track of the original song speed / track length).  easiest
to just use defaults and let advanced users modify settings
in the track editor.